Refactor: 패키지명 변경 (entity -> domain)
