+++ src/main/java/domain/Car.java
@@ -1,4 +1,4 @@
-package entity;
+package domain;
 
 public class Car {
