+++ src/main/java/domain/Racing.java
@@ -1,4 +1,4 @@
-package entity;
+package domain;
 
 import java.util.ArrayList;
 import java.util.List;
