+++ src/test/java/domain/CarTest.java
@@ -1,10 +1,8 @@
-package entity;
+package domain;
 
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class CarTest {
