+++ src/test/java/domain/RacingTest.java
@@ -1,4 +1,4 @@
-package entity;
+package domain;
 
 import org.assertj.core.api.Assertions;
 
@@ -6,8 +6,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class RacingTest {
